+++ apps/api/src/modules/products/products.module.ts
@@ -5,6 +5,8 @@ import { AccountsController } from "./accounts.controller";
 import { AccountsService } from "./accounts.service";
 import { CounterpartiesController } from "./counterparties.controller";
 import { CounterpartiesService } from "./counterparties.service";
+import { ProductCategoriesController } from "./product-categories.controller";
+import { ProductCategoriesService } from "./product-categories.service";
 import { ProductsController } from "./products.controller";
 import { ProductsService } from "./products.service";
 import { WarehousesController } from "./warehouses.controller";
@@ -12,6 +14,7 @@ import { WarehousesService } from "./warehouses.service";
 
 @Module({
   controllers: [
+    ProductCategoriesController,
     ProductsController,
     WarehousesController,
     CounterpartiesController,
@@ -19,12 +22,14 @@ import { WarehousesService } from "./warehouses.service";
   ],
   providers: [
     PrismaService,
+    ProductCategoriesService,
     ProductsService,
     WarehousesService,
     CounterpartiesService,
     AccountsService
   ],
   exports: [
+    ProductCategoriesService,
     ProductsService,
     WarehousesService,
     CounterpartiesService,